Create OnChainAllowList.sol
Creates an allowlist of token addresses OnChain.  When a bet is created, the main contract will need to check if the SkinToken is in the allowlist.  If it is not Allowed then the bet will be rejected.  This is to protect against allowing rebase tokens that could lead to irrevocable loss of funds.